requirements.txt
alembic==1.12.0
blinker==1.6.3
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.2.5
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-Navigation==0.2.0
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.8.0
Werkzeug==2.3.7
WTForms==3.0.1
supabase
